Refactor: Simplify timestamp retrieval and improve guard statements across multiple components
- Replaced asynchronous clock reference for timestamp retrieval with direct Date instantiation in SchemaEvolutionManager.
- Updated guard statements in ForeignKeyManager, BTreeIndexManager, SQLQueryExecutorManager, and ShardingManager to use `let` instead of `var` for better clarity and immutability.
- Enhanced PointInTimeRecoveryManager and VacuumManager by refining variable usage and ensuring consistent handling of optional values.
- Adjusted Logger to ensure line numbers are cast to UInt for type safety.

+++ Sources/ColibriCore/Utilities/Logger.swift
@@ -275,7 +275,7 @@ public actor Logger {
             message: message,
             file: file,
             function: function,
-            line: line,
+            line: UInt(line),
             metadata: metadata
         )
